Variable recording class
